+++ types.h
@@ -12,7 +12,7 @@ struct BitShare {
 // For INT_SUM, AND_OP, OR_OP
 struct IntShare {
     char pk[PK_LENGTH];
-    uint32_t val;
+    uint64_t val;
     char signature[PK_LENGTH];
 };
 
@@ -26,15 +26,15 @@ struct MaxShare {
 // For Var, Stddev
 struct VarShare {
     char pk[PK_LENGTH];
-    uint32_t val;
-    uint32_t val_squared;
+    uint64_t val;
+    uint64_t val_squared;
     char signature[PK_LENGTH];
 };
 
 struct LinRegShare {
     char pk[PK_LENGTH];
     uint32_t num_fields;
-    uint32_t* vals;
+    uint64_t* vals;
     char signature[PK_LENGTH];
 };
